Struct google_digitalassetlinks1::api::AssetlinkCheckCall

pub struct AssetlinkCheckCall<'a, C> where
    C: 'a,  { /* fields omitted */ }

Determines whether the specified (directional) relationship exists between the specified source and target assets. The relation describes the intent of the link between the two assets as claimed by the source asset. An example for such relationships is the delegation of privileges or permissions. This command is most often used by infrastructure systems to check preconditions for an action. For example, a client may want to know if it is OK to send a web URL to a particular mobile app instead. The client can check for the relevant asset link from the website to the mobile app to decide if the operation should be allowed. A note about security: if you specify a secure asset as the source, such as an HTTPS website or an Android app, the API will ensure that any statements used to generate the response have been made in a secure way by the owner of that asset. Conversely, if the source asset is an insecure HTTP website (that is, the URL starts with http:// instead of https://), the API cannot verify its statements securely, and it is not possible to ensure that the website's statements have not been altered by a third party. For more information, see the Digital Asset Links technical design specification.

A builder for the check method supported by a assetlink resource. It is not used directly, but through a AssetlinkMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.assetlinks().check()
             .target_web_site("amet")
             .target_android_app_package_name("duo")
             .target_android_app_certificate_sha256_fingerprint("ipsum")
             .source_web_site("sed")
             .source_android_app_package_name("ut")
             .source_android_app_certificate_sha256_fingerprint("gubergren")
             .relation("rebum.")
             .doit().await;

Implementations

impl<'a, C> AssetlinkCheckCall<'a, C> where
C: BorrowMut<Client<HttpsConnector<HttpConnector>, Body>>,

pub async fn doit(self) -> Result<(Response<Body>, CheckResponse)>

Perform the operation you have build so far.

pub fn target_web_site(self, new_value: &str) -> AssetlinkCheckCall<'a, C>

Web assets are identified by a URL that contains only the scheme, hostname and port parts. The format is http[s]://[:] Hostnames must be fully qualified: they must end in a single period ("."). Only the schemes "http" and "https" are currently allowed. Port numbers are given as a decimal number, and they must be omitted if the standard port numbers are used: 80 for http and 443 for https. We call this limited URL the "site". All URLs that share the same scheme, hostname and port are considered to be a part of the site and thus belong to the web asset. Example: the asset with the site https://www.google.com contains all these URLs: * https://www.google.com/ * https://www.google.com:443/ * https://www.google.com/foo * https://www.google.com/foo?bar * https://www.google.com/foo#bar * https://user@password:www.google.com/ But it does not contain these URLs: * http://www.google.com/ (wrong scheme) * https://google.com/ (hostname does not match) * https://www.google.com:444/ (port does not match) REQUIRED

Sets the target.web.site query property to the given value.

pub fn target_android_app_package_name(
self,
new_value: &str
) -> AssetlinkCheckCall<'a, C>

Android App assets are naturally identified by their Java package name. For example, the Google Maps app uses the package name com.google.android.apps.maps. REQUIRED

Sets the target.android app.package name query property to the given value.

pub fn target_android_app_certificate_sha256_fingerprint(
self,
new_value: &str
) -> AssetlinkCheckCall<'a, C>

The uppercase SHA-265 fingerprint of the certificate. From the PEM certificate, it can be acquired like this: $ keytool -printcert -file $CERTFILE | grep SHA256: SHA256: 14:6D:E9:83:C5:73:06:50:D8:EE:B9:95:2F:34:FC:64:16:A0:83: \ 42:E6:1D:BE:A8:8A:04:96:B2:3F:CF:44:E5 or like this: $ openssl x509 -in $CERTFILE -noout -fingerprint -sha256 SHA256 Fingerprint=14:6D:E9:83:C5:73:06:50:D8:EE:B9:95:2F:34:FC:64: \ 16:A0:83:42:E6:1D:BE:A8:8A:04:96:B2:3F:CF:44:E5 In this example, the contents of this field would be 14:6D:E9:83:C5:73: 06:50:D8:EE:B9:95:2F:34:FC:64:16:A0:83:42:E6:1D:BE:A8:8A:04:96:B2:3F:CF: 44:E5. If these tools are not available to you, you can convert the PEM certificate into the DER format, compute the SHA-256 hash of that string and represent the result as a hexstring (that is, uppercase hexadecimal representations of each octet, separated by colons).

Sets the target.android app.certificate.sha256 fingerprint query property to the given value.

pub fn source_web_site(self, new_value: &str) -> AssetlinkCheckCall<'a, C>

Web assets are identified by a URL that contains only the scheme, hostname and port parts. The format is http[s]://[:] Hostnames must be fully qualified: they must end in a single period ("."). Only the schemes "http" and "https" are currently allowed. Port numbers are given as a decimal number, and they must be omitted if the standard port numbers are used: 80 for http and 443 for https. We call this limited URL the "site". All URLs that share the same scheme, hostname and port are considered to be a part of the site and thus belong to the web asset. Example: the asset with the site https://www.google.com contains all these URLs: * https://www.google.com/ * https://www.google.com:443/ * https://www.google.com/foo * https://www.google.com/foo?bar * https://www.google.com/foo#bar * https://user@password:www.google.com/ But it does not contain these URLs: * http://www.google.com/ (wrong scheme) * https://google.com/ (hostname does not match) * https://www.google.com:444/ (port does not match) REQUIRED

Sets the source.web.site query property to the given value.

pub fn source_android_app_package_name(
self,
new_value: &str
) -> AssetlinkCheckCall<'a, C>

Android App assets are naturally identified by their Java package name. For example, the Google Maps app uses the package name com.google.android.apps.maps. REQUIRED

Sets the source.android app.package name query property to the given value.

pub fn source_android_app_certificate_sha256_fingerprint(
self,
new_value: &str
) -> AssetlinkCheckCall<'a, C>

The uppercase SHA-265 fingerprint of the certificate. From the PEM certificate, it can be acquired like this: $ keytool -printcert -file $CERTFILE | grep SHA256: SHA256: 14:6D:E9:83:C5:73:06:50:D8:EE:B9:95:2F:34:FC:64:16:A0:83: \ 42:E6:1D:BE:A8:8A:04:96:B2:3F:CF:44:E5 or like this: $ openssl x509 -in $CERTFILE -noout -fingerprint -sha256 SHA256 Fingerprint=14:6D:E9:83:C5:73:06:50:D8:EE:B9:95:2F:34:FC:64: \ 16:A0:83:42:E6:1D:BE:A8:8A:04:96:B2:3F:CF:44:E5 In this example, the contents of this field would be 14:6D:E9:83:C5:73: 06:50:D8:EE:B9:95:2F:34:FC:64:16:A0:83:42:E6:1D:BE:A8:8A:04:96:B2:3F:CF: 44:E5. If these tools are not available to you, you can convert the PEM certificate into the DER format, compute the SHA-256 hash of that string and represent the result as a hexstring (that is, uppercase hexadecimal representations of each octet, separated by colons).

Sets the source.android app.certificate.sha256 fingerprint query property to the given value.

pub fn relation(self, new_value: &str) -> AssetlinkCheckCall<'a, C>

Query string for the relation. We identify relations with strings of the format /, where must be one of a set of pre-defined purpose categories, and is a free-form lowercase alphanumeric string that describes the specific use case of the statement. Refer to our API documentation for the current list of supported relations. For a query to match an asset link, both the query's and the asset link's relation strings must match exactly. Example: A query with relation delegate_permission/common.handle_all_urls matches an asset link with relation delegate_permission/common.handle_all_urls.

Sets the relation query property to the given value.

pub fn delegate(
self,
new_value: &'a mut dyn Delegate
) -> AssetlinkCheckCall<'a, C>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

pub fn param<T>(self, name: T, value: T) -> AssetlinkCheckCall<'a, C> where
T: AsRef<str>,

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known parameters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

$.xgafv (query-string) - V1 error format.

access_token (query-string) - OAuth access token.

alt (query-string) - Data format for response.

callback (query-string) - JSONP

fields (query-string) - Selector specifying which fields to include in a partial response.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

oauth_token (query-string) - OAuth 2.0 token for the current user.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.

uploadType (query-string) - Legacy upload protocol for media (e.g. "media", "multipart").

upload_protocol (query-string) - Upload protocol for media (e.g. "raw", "multipart").