Struct google_digitalassetlinks1::api::Digitalassetlinks

pub struct Digitalassetlinks<C> { /* fields omitted */ }

Central instance to access all Digitalassetlinks related resource activities

Examples

Instantiate a new hub

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_digitalassetlinks1 as digitalassetlinks1;
use digitalassetlinks1::{Result, Error};
use std::default::Default;
use oauth2;
use digitalassetlinks1::Digitalassetlinks;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: oauth2::ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = Digitalassetlinks::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.assetlinks().check()
             .target_web_site("Lorem")
             .target_android_app_package_name("gubergren")
             .target_android_app_certificate_sha256_fingerprint("eos")
             .source_web_site("dolor")
             .source_android_app_package_name("ea")
             .source_android_app_certificate_sha256_fingerprint("ipsum")
             .relation("invidunt")
             .doit().await;
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::Io(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Implementations

impl<'a, C> Digitalassetlinks<C> where
C: BorrowMut<Client<HttpsConnector<HttpConnector>, Body>>,

pub fn new(
client: C,
authenticator: Authenticator<HttpsConnector<HttpConnector>>
) -> Digitalassetlinks<C>

pub fn assetlinks(&'a self) -> AssetlinkMethods<'a, C>

pub fn statements(&'a self) -> StatementMethods<'a, C>

pub fn user_agent(&mut self, agent_name: String) -> String

Set the user-agent header field to use in all requests to the server. It defaults to google-api-rust-client/2.0.0.

Returns the previously set user-agent.

pub fn base_url(&mut self, new_base_url: String) -> String

Set the base url to use in all requests to the server. It defaults to https://digitalassetlinks.googleapis.com/.

Returns the previously set base url.

pub fn root_url(&mut self, new_root_url: String) -> String

Set the root url to use in all requests to the server. It defaults to https://digitalassetlinks.googleapis.com/.

Returns the previously set root url.